you can now see your saved project. still needs some tweeks but works for the most part, saving to local storage works somewhat.
it saves to local storage successfully but editing the project and saving the changes again saves a whole new project. so if you create a prtoject and save it, it shows one project but if you go back and edit that project then save it again it will show up as 2 projects instead of one. one is older and the other is newer, same title unless you changed title.